ansibleinviewer/tests/requirements.txt
parameterized==0.7.0
pytest==5.2.2
PyYAML>=4.0
